* Fixed compiling errors. * Reverted the "fix" for premium issue, that change won't have any effect at all because the denominator is already a double.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4215 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ iologindata.cpp
@@ -335,7 +335,7 @@ void IOLoginData::removePremium(Account account)
 	uint64_t timeNow = time(NULL);
 	if(account.premiumDays > 0 && account.premiumDays < 65535)
 	{
-		uint32_t days = (uint32_t)std::ceil((double)(timeNow - account.lastDay) / 86400.);
+		uint32_t days = (uint32_t)std::ceil((timeNow - account.lastDay) / 86400.);
 		if(days > 0)
 		{
 			if(account.premiumDays >= days)
